+++ src/components/App.js
@@ -1,6 +1,6 @@
 import React from "react";
-import NavBar from "./containers/NavBar";
-import NavSub from "./containers/NavSub";
+import NavBar from "../containers/NavBar";
+import NavSub from "../containers/NavSub";
 
 const App = () => (
   <>
